chexnet-service/requirements.txt
flask==3.0.0
torch==2.1.2+cpu
torchvision==0.16.2+cpu
numpy==1.26.2
Pillow==10.1.0
matplotlib==3.8.2
--extra-index-url https://download.pytorch.org/whl/cpu